+++ podcasts/End of Year/Stories/TopOnePodcastStory.swift
@@ -2,7 +2,7 @@ import SwiftUI
 import PocketCastsServer
 import PocketCastsDataModel
 
-struct TopOnePodcastStory: StoryView {
+struct TopOnePodcastStory: ShareableStory {
     var duration: TimeInterval = 5.seconds
 
     let identifier: String = "top_one_podcast"
@@ -18,62 +18,23 @@ struct TopOnePodcastStory: StoryView {
     }
 
     var body: some View {
+        let podcast = topPodcast.podcast
         GeometryReader { geometry in
-            ZStack {
-                DynamicBackgroundView(podcast: topPodcast.podcast)
+            PodcastCoverContainer(geometry: geometry) {
+                PodcastStackView(podcasts: [podcast], geometry: geometry)
 
-                VStack {
-                    VStack {
-                        ZStack {
-                            let size = geometry.size.width * 0.60
-                            Rectangle().frame(width: size, height: size)
-                                .foregroundColor(ColorManager.darkThemeTintForPodcast(topPodcast.podcast).color)
-                                .modifier(BigCoverShadow())
-                                .modifier(PodcastCoverPerspective())
-                                .padding(.top, (size * 0.6))
+                StoryLabelContainer(geometry: geometry) {
+                    let title = podcast.title ?? ""
+                    let author = podcast.author ?? ""
+                    StoryLabel(L10n.eoyStoryTopPodcast("\n" + title, author), highlighting: [title, author], for: .title)
 
-                            Rectangle().frame(width: size, height: size)
-                                .foregroundColor(ColorManager.lightThemeTintForPodcast(topPodcast.podcast).color)
-                                .modifier(BigCoverShadow())
-                                .modifier(PodcastCoverPerspective())
-                                .padding(.top, (size * 0.30))
-
-                            PodcastCover(podcastUuid: topPodcast.podcast.uuid, big: true)
-                                .frame(width: size, height: size)
-                                .modifier(PodcastCoverPerspective())
-                        }
-
-                        Text(L10n.eoyStoryTopPodcast(topPodcast.podcast.title ?? "", topPodcast.podcast.author ?? ""))
-                            .foregroundColor(.white)
-                            .font(.system(size: 25, weight: .heavy))
-                            .foregroundColor(.white)
-                            .multilineTextAlignment(.center)
-                            .frame(maxHeight: geometry.size.height * 0.12)
-                            .minimumScaleFactor(0.01)
-                        Text(L10n.eoyStoryTopPodcastSubtitle(topPodcast.numberOfPlayedEpisodes, topPodcast.totalPlayedTime.localizedTimeDescription ?? ""))
-                            .font(.system(size: 15, weight: .bold))
-                            .foregroundColor(.white)
-                            .multilineTextAlignment(.center)
-                            .frame(maxHeight: geometry.size.height * 0.07)
-                            .minimumScaleFactor(0.01)
-                            .opacity(0.8)
-                    }
-                    .padding(.leading, 40)
-                    .padding(.trailing, 40)
+                    let time = topPodcast.totalPlayedTime.storyTimeDescription
+                    let count = L10n.eoyStoryListenedToEpisodesText(topPodcast.numberOfPlayedEpisodes)
+                    StoryLabel(L10n.eoyStoryTopPodcastSubtitle(topPodcast.numberOfPlayedEpisodes, time), highlighting: [time, count], for: .subtitle)
+                        .opacity(0.8)
                 }
             }
-            .padding(.top, -(geometry.size.height * 0.15))
-
-            VStack {
-                Spacer()
-                HStack {
-                    Spacer()
-                    Image("logo_white")
-                        .padding(.bottom, 40)
-                    Spacer()
-                }
-            }
-        }
+        }.background(DynamicBackgroundView(podcast: podcast))
     }
 
     func onAppear() {